+++ CHANGELOG.md
@@ -57,6 +57,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - portclippy, portscan: Look up `opt_USES_OFF` and `opt_VARS_OFF` too
 - portedit set-version: Deal with `PORTREVISION?=` and reset it to 0
 - portedit, portfmt: Ignore `-i` when `-D` was specified
+- portedit bump-epoch: Reset `PORTREVISION` on `PORTEPOCH` bump
 
 ## [g20210321] - 2021-03-21